:recycle: Split the route.rs into different files
