+++ src/lib.rs
@@ -3,5 +3,8 @@ extern crate diesel;
 
 pub mod fairings;
 pub mod models;
-pub mod routes;
+pub mod routes {
+    pub mod health;
+    pub mod version;
+}
 pub mod schema;
