+++ src/main.rs
@@ -71,7 +71,7 @@ fn setup_logging(logging_level: LevelFilter) {
 async fn main() {
     use log::{debug, error, info};
     use minne_backend::fairings::{BackendConfiguration, MinneDatabaseConnection, NoCacheFairing};
-    use minne_backend::routes::{check_backend_health, get_backend_version};
+    use minne_backend::routes::{health::check_backend_health, version::get_backend_version};
     use rocket::config::{Shutdown, Sig};
     use rocket::figment::{
         util::map,
